+++ lib/src/widgets/indicators/loader_indicator.dart
@@ -11,13 +11,13 @@ class LoaderIndicator extends StatelessWidget {
     return Container(
       padding: const EdgeInsets.all(20),
       decoration: BoxDecoration(
-        color: theme.colors.playPauseBackground.withAlpha(100),
+        color: theme.colors.controlButtonBackground.withAlpha(100),
         borderRadius: BorderRadius.circular(200),
       ),
       child: CircularProgressIndicator(
         strokeWidth: 3,
         valueColor: AlwaysStoppedAnimation<Color>(
-          theme.colors.playPauseIcon ?? theme.colors.icon ?? Colors.white,
+          theme.colors.controlButtonIcon ?? theme.colors.icon ?? Colors.white,
         ),
       ),
     );